# docs/swagger.yaml
basePath: /api
definitions:
  api_err.ApiErr:
    properties:
      msg:
        type: string
      typ:
        $ref: '#/definitions/api_err.ApiErrTyp'
    type: object
  api_err.ApiErrTyp:
    enum:
    - not_found
    - bad_input
    - forbidden
    - unauthorized
    - already_claimed
    - pending_claim_exists
    - username_invalid
    - username_taken
    - report_target_missing
    - report_locked
    - duplicate_open_report
    - text_too_long
    - unknown
    type: string
    x-enum-varnames:
    - AetNotFound
    - AetBadInput
    - AetForbidden
    - AetUnauthorized
    - AetAlreadyClaimed
    - AetPendingClaimExists
    - AetUsernameInvalid
    - AetUsernameTaken
    - AetReportTargetMissing
    - AetReportLocked
    - AetDuplicateOpenReport
    - AetTextTooLong
    - AetUnknown
...
  models.GeoPoint:
    properties:
      coordinates:
        items:
          type: number
        type: array
      type:
        type: string
    type: object
  models.HourRange:
    properties:
      close:
        description: '"22:00"'
        type: string
      open:
        description: '"09:00"'
        type: string
    type: object
  models.I18nText:
    properties:
      en:
        type: string
      uz:
        type: string
    type: object
  models.PublicUser:
    properties:
      avatar_key:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      username:
        type: string
    type: object
...
  models.Status:
    enum:
    - -10
    - 0
    - 10
    type: integer
    x-enum-varnames:
    - StatusRejected
    - StatusPending
    - StatusApproved
  models.UserMini:
    properties:
      avatar_key:
        type: string
      id:
        type: string
      name:
        type: string
      username:
        type: string
    type: object
  models.WeeklyHours:
    properties:
      fri:
        items:
          $ref: '#/definitions/models.HourRange'
        type: array
      mon:
        items:
          $ref: '#/definitions/models.HourRange'
        type: array
      sat:
        items:
          $ref: '#/definitions/models.HourRange'
        type: array
      sun:
        items:
          $ref: '#/definitions/models.HourRange'
        type: array
      thu:
        items:
          $ref: '#/definitions/models.HourRange'
        type: array
      tue:
        items:
          $ref: '#/definitions/models.HourRange'
        type: array
      wed:
        items:
          $ref: '#/definitions/models.HourRange'
        type: array
    type: object
...
  services.BookmarkView:
    properties:
      created_at:
        type: string
      id:
        type: string
      place:
        $ref: '#/definitions/services.PlaceView'
      place_id:
        type: string
      user_id:
        type: string
    type: object
  services.ClaimView:
    properties:
      created_at:
        type: string
      id:
        type: string
      note:
        type: string
      phone:
        type: string
      place_id:
        type: string
      reviewed_by:
        type: string
      status:
        $ref: '#/definitions/models.Status'
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.UserMini'
      user_id:
        type: string
    type: object
...
  services.CreatePlaceInput:
    properties:
      address:
        $ref: '#/definitions/models.I18nText'
      category_id:
        type: string
      description:
        $ref: '#/definitions/models.I18nText'
      images:
        items:
          type: string
        type: array
      lat:
        type: number
      logo_key:
        type: string
      lon:
        type: number
      name:
        type: string
      phone:
        type: string
      weekly_hours:
        $ref: '#/definitions/models.WeeklyHours'
    type: object
...
  services.EditPlaceInput:
    properties:
      description:
        $ref: '#/definitions/models.I18nText'
      images:
        items:
          type: string
        type: array
      logo_key:
        type: string
      phone:
        type: string
      weekly_hours:
        $ref: '#/definitions/models.WeeklyHours'
    type: object
...
  services.MeView:
    properties:
      avatar_key:
        type: string
      blocked:
        type: boolean
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      owns_place:
        type: boolean
      phone:
        type: string
      username:
        type: string
    type: object
...
  services.PlaceView:
    properties:
      address:
        $ref: '#/definitions/models.I18nText'
      atc_id:
        type: string
      avg_rating:
        type: number
      category_id:
        type: string
      category_name:
        $ref: '#/definitions/models.I18nText'
      claimed_by:
        type: string
      claimed_by_user:
        $ref: '#/definitions/models.UserMini'
      created_at:
        type: string
      created_by:
        type: string
      created_by_user:
        $ref: '#/definitions/models.UserMini'
      description:
        $ref: '#/definitions/models.I18nText'
      id:
        type: string
      images:
        items:
          type: string
        type: array
      is_open:
        type: boolean
      lat:
        type: number
      location:
        $ref: '#/definitions/models.GeoPoint'
      logo_key:
        type: string
      lon:
        type: number
      name:
        type: string
      phone:
        type: string
      review_count:
        type: integer
      slug:
        type: string
      status:
        $ref: '#/definitions/models.Status'
      updated_at:
        type: string
      weekly_hours:
        $ref: '#/definitions/models.WeeklyHours'
    type: object
...
  services.SubmitClaimInput:
    properties:
      note:
        type: string
      phone:
        type: string
      place_id:
        type: string
    type: object
...
  services.VerifyCodeInput:
    properties:
      code:
        type: string
    type: object
  services.VerifyCodeOutput:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/models.PublicUser'
    type: object
host: localhost:8080
info:
  contact: {}
  description: S101 backend API
  title: S101 API
  version: "1.0"
paths:
...
  /auth/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.MeView'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_err.ApiErr'
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - auth
  /auth/verify-code:
    post:
      consumes:
      - application/json
      parameters:
      - description: OTP code
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/services.VerifyCodeInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.VerifyCodeOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_err.ApiErr'
      summary: Verify OTP code and get token
      tags:
      - auth
  /bookmarks:
    get:
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_err.ApiErr'
      security:
      - BearerAuth: []
      summary: List own bookmarks
      tags:
      - bookmarks
  /bookmarks/{placeId}:
    delete:
      parameters:
      - description: Place ID
        in: path
        name: placeId
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_err.ApiErr'
      security:
      - BearerAuth: []
      summary: Remove bookmark
      tags:
      - bookmarks
    post:
      parameters:
      - description: Place ID
        in: path
        name: placeId
        required: true
        type: string
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/services.BookmarkView'
        "208":
          description: Already Reported
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_err.ApiErr'
      security:
      - BearerAuth: []
      summary: Add bookmark
      tags:
      - bookmarks
...
  /claims:
    post:
      consumes:
      - application/json
      parameters:
      - description: Claim data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/services.SubmitClaimInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/services.ClaimView'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_err.ApiErr'
      security:
      - BearerAuth: []
      summary: Submit a claim for a place
      tags:
      - claims
  /claims/mine:
    get:
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_err.ApiErr'
      security:
      - BearerAuth: []
      summary: List own claims
      tags:
      - claims
  /files/upload:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: File to upload
        in: formData
        name: file
        required: true
        type: file
      - description: Usage hint (e.g. avatar, place_logo)
        in: formData
        name: usage
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            properties:
              file_id:
                type: string
              key:
                type: string
              url:
                type: string
              usage:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_err.ApiErr'
      security:
      - BearerAuth: []
      summary: Upload a file
      tags:
      - files
  /places:
    get:
      parameters:
      - description: Search query
        in: query
        name: query
        type: string
      - description: 'Sort: top|recent|nearest'
        in: query
        name: sort
        type: string
      - description: Category UUID
        in: query
        name: category_id
        type: string
      - description: lat,lon for nearest sort
        in: query
        name: near
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: List places
      tags:
      - places
  /places/{id}:
    get:
      parameters:
      - description: Place UUID or slug
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.PlaceView'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api_err.ApiErr'
      summary: Get place by ID or slug
      tags:
      - places
    put:
      consumes:
      - application/json
      parameters:
      - description: Place UUID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/services.EditPlaceInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.PlaceView'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api_err.ApiErr'
      security:
      - BearerAuth: []
      summary: Edit own place (claimant only)
      tags:
      - places
...
  /places/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Place data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/services.CreatePlaceInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/services.PlaceView'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_err.ApiErr'
      security:
      - BearerAuth: []
      summary: Create a place
      tags:
      - places
...
securityDefinitions:
  BearerAuth:
    description: 'Format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"